test: remove unused rif-relay-common dependency remove TestUtils.ts and include the missing functions in utils.ts
