+++ package.json
@@ -46,7 +46,6 @@
   "devDependencies": {
     "@openeth/truffle-typings": "0.0.6",
     "@openzeppelin/test-helpers": "^0.5.15",
-    "@rsksmart/rif-relay-common": "0.0.2-alpha.3",
     "@truffle/contract-schema": "3.3.2",
     "@truffle/hdwallet-provider": "1.2.4",
     "@trufflesuite/web3-provider-engine": "^15.0.13-1",
